pyproject.toml
[project]
name = "paper-writing-agent"
version = "0.1.0"
description = "Multi-agent academic paper writing system using CrewAI + ZhipuAI GLM"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "tools", "agents", "crews"]
